chore: move getFluxExample to utils folder
