+++ src/timeMachine/components/TimeMachineFluxEditor.tsx
@@ -19,6 +19,7 @@ import {getActiveQuery, getActiveTimeMachine} from 'src/timeMachine/selectors'
 import {
   functionRequiresNewLine,
   generateImport,
+  getFluxExample,
 } from 'src/timeMachine/utils/insertFunction'
 import {event} from 'src/cloud/utils/reporting'
 import {isFlagEnabled} from 'src/shared/utils/featureFlag'
@@ -26,7 +27,6 @@ import {isFlagEnabled} from 'src/shared/utils/featureFlag'
 // Types
 import {FluxToolbarFunction, EditorType} from 'src/types'
 import * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api'
-
 
 const FluxEditor = lazy(() => import('src/shared/components/FluxMonacoEditor'))
 
@@ -125,117 +125,13 @@ const TimeMachineFluxEditor: FC = () => {
     return {text, range}
   }
 
-  // dynamic flux example parser function
-  const getFluxExample = func => {
-
-    const {name, fluxType} = func
-    let signature
-    // aggregate.rate
-    // "fluxType": "(<-tables:stream[A], every:duration, ?groupColumns:[string], ?unit:duration) => stream[B] where A: Record, B: Record"
-
-
-    // get copy of fluxtype signature before arrow sign
-    const arrowSign = fluxType.lastIndexOf('=') // using lastIndexOf because of edge cases like array.filter
-    const fluxParam = fluxType.slice(0, arrowSign) // fluxParam = (<-tables:stream[A], every:duration, ?groupColumns:[string], ?unit:duration)
-
-    // pull out all parameters found in between paranethesis 
-    const openingParenthesis = fluxParam.indexOf('(')
-    const closingParenthesis = fluxParam.lastIndexOf(')')
-    const parametersAsOneSentence = fluxParam
-      .substring(openingParenthesis + 1, closingParenthesis)
-      .replace(/\s/g, '')
-
-    console.log('all parameters ', parametersAsOneSentence)
-    
-    // parametersAsOneSentence = <-tables:stream[A],every:duration,?groupColumns:[string],?unit:duration
-
-    // sparate parametersAsOneSentence into individual Parameters 
-    // some flux parameters cant be sperated by commas because some stand alone params contain commas. ex below 
-    // (t:A, ?location:{zone:string, offset:duration}) => int where A: Timeable"
-    // below code separates them by keeping track of opening and closing brackets 
-
-    const individualParams = []
-    const stack = [] 
-    const brackets = {
-        '(': ')',
-        '[': ']',
-        '{': '}'
-      }
-    let param = ''
-    
-    for (let i = 0; i < parametersAsOneSentence.length; i++) {
-      if (parametersAsOneSentence[i] === ',' && !stack.length) { // case: params with no brackets in them
-        if (!param.startsWith('?') && !param.startsWith('<')) { // checks if param is optional. if not, push to individialParams array
-          individualParams.push(param)
-          param = ''
-          i++
-        }
-      }
-      param += parametersAsOneSentence[i] 
-
-      if (brackets.hasOwnProperty(parametersAsOneSentence[i]) ) { // if element is opening bracket 
-        stack.push(parametersAsOneSentence[i])
-      } 
-      if (Object.values(brackets).includes(parametersAsOneSentence[i])) { // if element is closing bracket 
-        const closingBracket = stack.pop()
-        // if element is the correct closing bracket AND stack is empty 
-        // AND the next element is a comma, we have a complete parameter. 
-        if (parametersAsOneSentence[i] === brackets[closingBracket] && !stack.length && parametersAsOneSentence[i + 1] === ',') { 
-          if (!param.startsWith('?') && !param.startsWith('<')) { // checks if param is optional
-            individualParams.push(param)
-            param = ''
-            i++
-          } else { // we have an optional param. set param to empty
-            param = ''
-            i++
-          }
-        }
-      }
-      if (i === parametersAsOneSentence.length - 1) { // end of iteration
-        if (!param.startsWith('?') && !param.startsWith('<')) { 
-          individualParams.push(param)
-        }
-      }
-    }
-
-    // at this point, individualParams array should have all required parameters to parse a signature 
-    // temp solution for missing parameter type info 
-    // manually updating the value of parameter's property 
-
-    /* individualParams = [
-      'default:A',
-      'dict:[B:A]',
-      'key:B'
-      ]
-    */ 
-    individualParams.forEach((element, index) => {
-     
-      // remove parameter placeholders 
-      const emptyPlaceholder = element.split(':')[0] + ': ' // dict: 
-      individualParams[index] = emptyPlaceholder
-
-    })
-    console.log('required params ', individualParams)
-     /* individualParams = [
-      'default: ',
-      'dict: ',
-      'key: '
-      ]
-    */ 
-
-    // join the individual params to create the signature
-    signature = `${func.package}.${name}` + `(` + individualParams.join(', ') + `)`
-
-    // add example property to flux function object
-    return {...func, example: signature}
-
-  }
-
   const handleInsertFluxFunction = (func: FluxToolbarFunction): void => {
     if (!editorInstance) {
       return
     }
-    const {text, range} = getFluxTextAndRange( isFlagEnabled('fluxDynamicDocs') ? getFluxExample(func) : func)
+    const {text, range} = getFluxTextAndRange(
+      isFlagEnabled('fluxDynamicDocs') ? getFluxExample(func) : func
+    )
 
     const edits = [
       {
